% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{Alelager}
\alias{Alelager}
\title{The Alelager data set.}
\format{A data frame with 31 rows and 4 variables.
\describe{
  \item{ID}{variable}
  \item{Type}{variable}
  \item{Alcohol}{variable}
  \item{Calories}{variable}
}}
\source{
\url{http://people.kzoo.edu/enordmoe/math365/data/}
}
\usage{
Alelager
}
\description{
The description of the Alelager data set.
}
\keyword{datasets}
