% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{AMPMFlightDelays}
\alias{AMPMFlightDelays}
\title{The AMPMFlightDelays data set.}
\format{A data frame with 1671 rows and 10 variables.
\describe{
  \item{ID}{variable}
  \item{Carrier}{variable}
  \item{FlightNo}{variable}
  \item{Destination}{variable}
  \item{DepartTime}{variable}
  \item{Day}{variable}
  \item{Month}{variable}
  \item{FlightLength}{variable}
  \item{Delay}{variable}
  \item{Delayed30}{variable}
}}
\source{
\url{http://people.kzoo.edu/enordmoe/math365/data/}
}
\usage{
AMPMFlightDelays
}
\description{
The description of the AMPMFlightDelays data set.
}
\keyword{datasets}
