% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{Bangladesh}
\alias{Bangladesh}
\title{The Bangladesh data set.}
\format{A data frame with 271 rows and 3 variables.
\describe{
  \item{Arsenic}{variable}
  \item{Chlorine}{variable}
  \item{Cobalt}{variable}
}}
\source{
\url{http://people.kzoo.edu/enordmoe/math365/data/}
}
\usage{
Bangladesh
}
\description{
The description of the Bangladesh data set.
}
\keyword{datasets}
