% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{Beerwings}
\alias{Beerwings}
\title{The Beerwings data set.}
\format{A data frame with 30 rows and 4 variables.
\describe{
  \item{ID}{variable}
  \item{Hotwings}{variable}
  \item{Beer}{variable}
  \item{Gender}{variable}
}}
\source{
\url{http://people.kzoo.edu/enordmoe/math365/data/}
}
\usage{
Beerwings
}
\description{
The description of the Beerwings data set.
}
\keyword{datasets}
