% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{BookPrices}
\alias{BookPrices}
\title{The BookPrices data set.}
\format{A data frame with 44 rows and 3 variables.
\describe{
  \item{Subject}{variable}
  \item{Area}{variable}
  \item{Price}{variable}
}}
\source{
\url{http://people.kzoo.edu/enordmoe/math365/data/}
}
\usage{
BookPrices
}
\description{
The description of the BookPrices data set.
}
\keyword{datasets}
