% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{BootExerciseA}
\alias{BootExerciseA}
\title{The BootExerciseA data set.}
\format{A data frame with 21 rows and 2 variables.
\describe{
  \item{ID}{variable}
  \item{X}{variable}
}}
\source{
\url{http://people.kzoo.edu/enordmoe/math365/data/}
}
\usage{
BootExerciseA
}
\description{
The description of the BootExerciseA data set.
}
\keyword{datasets}
