% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{Bushmeat}
\alias{Bushmeat}
\title{The Bushmeat data set.}
\format{A data frame with 30 rows and 3 variables.
\describe{
  \item{Fish}{variable}
  \item{Biomass}{variable}
  \item{Year}{variable}
}}
\source{
\url{http://people.kzoo.edu/enordmoe/math365/data/}
}
\usage{
Bushmeat
}
\description{
The description of the Bushmeat data set.
}
\keyword{datasets}
