% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{Cameras}
\alias{Cameras}
\title{The Cameras data set.}
\format{A data frame with 22 rows and 4 variables.
\describe{
  \item{ID}{variable}
  \item{Item}{variable}
  \item{JR}{variable}
  \item{BH}{variable}
}}
\source{
\url{http://people.kzoo.edu/enordmoe/math365/data/}
}
\usage{
Cameras
}
\description{
The description of the Cameras data set.
}
\keyword{datasets}
