% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{Cereals}
\alias{Cereals}
\title{The Cereals data set.}
\format{A data frame with 43 rows and 5 variables.
\describe{
  \item{ID}{variable}
  \item{Age}{variable}
  \item{Shelf}{variable}
  \item{Sodiumgram}{variable}
  \item{Proteingram}{variable}
}}
\source{
\url{http://people.kzoo.edu/enordmoe/math365/data/}
}
\usage{
Cereals
}
\description{
The description of the Cereals data set.
}
\keyword{datasets}
