% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{Challenger}
\alias{Challenger}
\title{The Challenger data set.}
\format{A data frame with 23 rows and 3 variables.
\describe{
  \item{Date}{variable}
  \item{Temperature}{variable}
  \item{Incident}{variable}
}}
\source{
\url{http://people.kzoo.edu/enordmoe/math365/data/}
}
\usage{
Challenger
}
\description{
The description of the Challenger data set.
}
\keyword{datasets}
