% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{cocaine}
\alias{cocaine}
\title{The cocaine data set.}
\format{A data frame with 48 rows and 2 variables.
\describe{
  \item{drug}{variable}
  \item{outcome}{variable}
}}
\source{
\url{http://people.kzoo.edu/enordmoe/math365/data/}
}
\usage{
cocaine
}
\description{
The description of the cocaine data set.
}
\keyword{datasets}
