% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{corrExerciseA}
\alias{corrExerciseA}
\title{The corrExerciseA data set.}
\format{A data frame with 200 rows and 3 variables.
\describe{
  \item{X}{variable}
  \item{Y}{variable}
  \item{Z}{variable}
}}
\source{
\url{http://people.kzoo.edu/enordmoe/math365/data/}
}
\usage{
corrExerciseA
}
\description{
The description of the corrExerciseA data set.
}
\keyword{datasets}
