% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{DeltaDelays}
\alias{DeltaDelays}
\title{The DeltaDelays data set.}
\format{A data frame with 3989 rows and 2 variables.
\describe{
  \item{Airport}{variable}
  \item{Delay}{variable}
}}
\source{
\url{http://people.kzoo.edu/enordmoe/math365/data/}
}
\usage{
DeltaDelays
}
\description{
The description of the DeltaDelays data set.
}
\keyword{datasets}
