% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{fasondata}
\alias{fasondata}
\title{The fasondata data set.}
\format{A data frame with 1078 rows and 2 variables.
\describe{
  \item{fheight}{variable}
  \item{sheight}{variable}
}}
\source{
\url{http://people.kzoo.edu/enordmoe/math365/data/}
}
\usage{
fasondata
}
\description{
The description of the fasondata data set.
}
\keyword{datasets}
