% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{Fatalities}
\alias{Fatalities}
\title{The Fatalities data set.}
\format{A data frame with 100 rows and 3 variables.
\describe{
  \item{ID}{variable}
  \item{Alcohol}{variable}
  \item{Age}{variable}
}}
\source{
\url{http://people.kzoo.edu/enordmoe/math365/data/}
}
\usage{
Fatalities
}
\description{
The description of the Fatalities data set.
}
\keyword{datasets}
