% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{FishMercury}
\alias{FishMercury}
\title{The FishMercury data set.}
\format{A data frame with 30 rows and 1 variables.
\describe{
  \item{Mercury}{variable}
}}
\source{
\url{http://people.kzoo.edu/enordmoe/math365/data/}
}
\usage{
FishMercury
}
\description{
The description of the FishMercury data set.
}
\keyword{datasets}
