% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{funnel}
\alias{funnel}
\title{The funnel data set.}
\format{A data frame with 18 rows and 3 variables.
\describe{
  \item{distance}{variable}
  \item{trial}{variable}
  \item{tswirl}{variable}
}}
\source{
\url{http://people.kzoo.edu/enordmoe/math365/data/}
}
\usage{
funnel
}
\description{
The description of the funnel data set.
}
\keyword{datasets}
