% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{Girls2004}
\alias{Girls2004}
\title{The Girls2004 data set.}
\format{A data frame with 80 rows and 6 variables.
\describe{
  \item{ID}{variable}
  \item{State}{variable}
  \item{MothersAge}{variable}
  \item{Smoker}{variable}
  \item{Weight}{variable}
  \item{Gestation}{variable}
}}
\source{
\url{http://people.kzoo.edu/enordmoe/math365/data/}
}
\usage{
Girls2004
}
\description{
The description of the Girls2004 data set.
}
\keyword{datasets}
