% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{GSS2002}
\alias{GSS2002}
\title{The GSS2002 data set.}
\format{A data frame with 2765 rows and 21 variables.
\describe{
  \item{ID}{variable}
  \item{Region}{variable}
  \item{Gender}{variable}
  \item{Race}{variable}
  \item{Education}{variable}
  \item{Marital}{variable}
  \item{Religion}{variable}
  \item{Happy}{variable}
  \item{Income}{variable}
  \item{PolParty}{variable}
  \item{Politics}{variable}
  \item{Marijuana}{variable}
  \item{DeathPenalty}{variable}
  \item{OwnGun}{variable}
  \item{GunLaw}{variable}
  \item{SpendMilitary}{variable}
  \item{SpendEduc}{variable}
  \item{SpendEnv}{variable}
  \item{SpendSci}{variable}
  \item{Pres00}{variable}
  \item{Postlife}{variable}
}}
\source{
\url{http://people.kzoo.edu/enordmoe/math365/data/}
}
\usage{
GSS2002
}
\description{
The description of the GSS2002 data set.
}
\keyword{datasets}
