% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{heightfoot}
\alias{heightfoot}
\title{The heightfoot data set.}
\format{A data frame with 42 rows and 2 variables.
\describe{
  \item{height}{variable}
  \item{foot}{variable}
}}
\source{
\url{http://people.kzoo.edu/enordmoe/math365/data/}
}
\usage{
heightfoot
}
\description{
The description of the heightfoot data set.
}
\keyword{datasets}
