% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{Hollywood}
\alias{Hollywood}
\title{The Hollywood data set.}
\format{A data frame with 136 rows and 3 variables.
\describe{
  \item{Movie}{variable}
  \item{Rotten}{variable}
  \item{Audience}{variable}
}}
\source{
\url{http://people.kzoo.edu/enordmoe/math365/data/}
}
\usage{
Hollywood
}
\description{
The description of the Hollywood data set.
}
\keyword{datasets}
