% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{IceCream}
\alias{IceCream}
\title{The IceCream data set.}
\format{A data frame with 39 rows and 7 variables.
\describe{
  \item{Brand}{variable}
  \item{VanillaCalories}{variable}
  \item{VanillaFat}{variable}
  \item{VanillaSugar}{variable}
  \item{ChocolateCalories}{variable}
  \item{ChocolateFat}{variable}
  \item{ChocolateSugar}{variable}
}}
\source{
\url{http://people.kzoo.edu/enordmoe/math365/data/}
}
\usage{
IceCream
}
\description{
The description of the IceCream data set.
}
\keyword{datasets}
