% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{ILBoys}
\alias{ILBoys}
\title{The ILBoys data set.}
\format{A data frame with 241 rows and 2 variables.
\describe{
  \item{MothersAge}{variable}
  \item{Weight}{variable}
}}
\source{
\url{http://people.kzoo.edu/enordmoe/math365/data/}
}
\usage{
ILBoys
}
\description{
The description of the ILBoys data set.
}
\keyword{datasets}
