% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{Illiteracy}
\alias{Illiteracy}
\title{The Illiteracy data set.}
\format{A data frame with 94 rows and 4 variables.
\describe{
  \item{ID}{variable}
  \item{Country}{variable}
  \item{Illit}{variable}
  \item{Births}{variable}
}}
\source{
\url{http://people.kzoo.edu/enordmoe/math365/data/}
}
\usage{
Illiteracy
}
\description{
The description of the Illiteracy data set.
}
\keyword{datasets}
