% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{Lottery}
\alias{Lottery}
\title{The Lottery data set.}
\format{A data frame with 500 rows and 1 variables.
\describe{
  \item{Win}{variable}
}}
\source{
\url{http://people.kzoo.edu/enordmoe/math365/data/}
}
\usage{
Lottery
}
\description{
The description of the Lottery data set.
}
\keyword{datasets}
