% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{math}
\alias{math}
\title{The math data set.}
\format{A data frame with 18 rows and 4 variables.
\describe{
  \item{group}{variable}
  \item{before}{variable}
  \item{after}{variable}
  \item{gain}{variable}
}}
\source{
\url{http://people.kzoo.edu/enordmoe/math365/data/}
}
\usage{
math
}
\description{
The description of the math data set.
}
\keyword{datasets}
