% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{Maunaloa}
\alias{Maunaloa}
\title{The Maunaloa data set.}
\format{A data frame with 21 rows and 3 variables.
\describe{
  \item{ID}{variable}
  \item{Year}{variable}
  \item{Level}{variable}
}}
\source{
\url{http://people.kzoo.edu/enordmoe/math365/data/}
}
\usage{
Maunaloa
}
\description{
The description of the Maunaloa data set.
}
\keyword{datasets}
