% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{Memory}
\alias{Memory}
\title{The Memory data set.}
\format{A data frame with 24 rows and 2 variables.
\describe{
  \item{Group}{variable}
  \item{Words}{variable}
}}
\source{
\url{http://people.kzoo.edu/enordmoe/math365/data/}
}
\usage{
Memory
}
\description{
The description of the Memory data set.
}
\keyword{datasets}
