% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{MnGroundwater}
\alias{MnGroundwater}
\title{The MnGroundwater data set.}
\format{A data frame with 895 rows and 10 variables.
\describe{
  \item{County}{variable}
  \item{Aquifer.Group}{variable}
  \item{Water.Level}{variable}
  \item{Alkalinity}{variable}
  \item{Aluminum}{variable}
  \item{Arsenic}{variable}
  \item{Chloride}{variable}
  \item{Lead}{variable}
  \item{pH}{variable}
  \item{Basin.Name}{variable}
}}
\source{
\url{http://people.kzoo.edu/enordmoe/math365/data/}
}
\usage{
MnGroundwater
}
\description{
The description of the MnGroundwater data set.
}
\keyword{datasets}
