% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{Mustang}
\alias{Mustang}
\title{The Mustang data set.}
\format{A data frame with 25 rows and 3 variables.
\describe{
  \item{Age}{variable}
  \item{Miles}{variable}
  \item{Price}{variable}
}}
\source{
\url{http://people.kzoo.edu/enordmoe/math365/data/}
}
\usage{
Mustang
}
\description{
The description of the Mustang data set.
}
\keyword{datasets}
