% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{NCBirths2004}
\alias{NCBirths2004}
\title{The NCBirths2004 data set.}
\format{A data frame with 1009 rows and 7 variables.
\describe{
  \item{ID}{variable}
  \item{MothersAge}{variable}
  \item{Tobacco}{variable}
  \item{Alcohol}{variable}
  \item{Gender}{variable}
  \item{Weight}{variable}
  \item{Gestation}{variable}
}}
\source{
\url{http://people.kzoo.edu/enordmoe/math365/data/}
}
\usage{
NCBirths2004
}
\description{
The description of the NCBirths2004 data set.
}
\keyword{datasets}
