% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{Phillies2009}
\alias{Phillies2009}
\title{The Phillies2009 data set.}
\format{A data frame with 162 rows and 7 variables.
\describe{
  \item{Date}{variable}
  \item{Location}{variable}
  \item{Outcome}{variable}
  \item{Hits}{variable}
  \item{Doubles}{variable}
  \item{HomeRuns}{variable}
  \item{StrikeOuts}{variable}
}}
\source{
\url{http://people.kzoo.edu/enordmoe/math365/data/}
}
\usage{
Phillies2009
}
\description{
The description of the Phillies2009 data set.
}
\keyword{datasets}
