% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{prob3p23data}
\alias{prob3p23data}
\title{The prob3p23data data set.}
\format{A data frame with 60 rows and 2 variables.
\describe{
  \item{X1}{variable}
  \item{x}{variable}
}}
\source{
\url{http://people.kzoo.edu/enordmoe/math365/data/}
}
\usage{
prob3p23data
}
\description{
The description of the prob3p23data data set.
}
\keyword{datasets}
