% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{Quakes}
\alias{Quakes}
\title{The Quakes data set.}
\format{A data frame with 805 rows and 2 variables.
\describe{
  \item{ID}{variable}
  \item{TimeDiff}{variable}
}}
\source{
\url{http://people.kzoo.edu/enordmoe/math365/data/}
}
\usage{
Quakes
}
\description{
The description of the Quakes data set.
}
\keyword{datasets}
