% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{Service}
\alias{Service}
\title{The Service data set.}
\format{A data frame with 174 rows and 2 variables.
\describe{
  \item{ID}{variable}
  \item{Times}{variable}
}}
\source{
\url{http://people.kzoo.edu/enordmoe/math365/data/}
}
\usage{
Service
}
\description{
The description of the Service data set.
}
\keyword{datasets}
