% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{Skating2010}
\alias{Skating2010}
\title{The Skating2010 data set.}
\format{A data frame with 24 rows and 5 variables.
\describe{
  \item{Name}{variable}
  \item{Country}{variable}
  \item{Short}{variable}
  \item{Free}{variable}
  \item{Total}{variable}
}}
\source{
\url{http://people.kzoo.edu/enordmoe/math365/data/}
}
\usage{
Skating2010
}
\description{
The description of the Skating2010 data set.
}
\keyword{datasets}
