% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{Sleep}
\alias{Sleep}
\title{The Sleep data set.}
\format{A data frame with 18 rows and 1 variables.
\describe{
  \item{Hours}{variable}
}}
\source{
\url{http://people.kzoo.edu/enordmoe/math365/data/}
}
\usage{
Sleep
}
\description{
The description of the Sleep data set.
}
\keyword{datasets}
