% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{soccer14}
\alias{soccer14}
\title{The soccer14 data set.}
\format{A data frame with 40 rows and 3 variables.
\describe{
  \item{League}{variable}
  \item{Team}{variable}
  \item{Goals}{variable}
}}
\source{
\url{http://people.kzoo.edu/enordmoe/math365/data/}
}
\usage{
soccer14
}
\description{
The description of the soccer14 data set.
}
\keyword{datasets}
