% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{Spruce}
\alias{Spruce}
\title{The Spruce data set.}
\format{A data frame with 72 rows and 9 variables.
\describe{
  \item{Tree}{variable}
  \item{Competition}{variable}
  \item{Fertilizer}{variable}
  \item{Height0}{variable}
  \item{Height5}{variable}
  \item{Diameter0}{variable}
  \item{Diameter5}{variable}
  \item{Ht.change}{variable}
  \item{Di.change}{variable}
}}
\source{
\url{http://people.kzoo.edu/enordmoe/math365/data/}
}
\usage{
Spruce
}
\description{
The description of the Spruce data set.
}
\keyword{datasets}
