% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{statedat}
\alias{statedat}
\title{The statedat data set.}
\format{A data frame with 50 rows and 3 variables.
\describe{
  \item{X1}{variable}
  \item{illit}{variable}
  \item{lifexp}{variable}
}}
\source{
\url{http://people.kzoo.edu/enordmoe/math365/data/}
}
\usage{
statedat
}
\description{
The description of the statedat data set.
}
\keyword{datasets}
