% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{stmeas}
\alias{stmeas}
\title{The stmeas data set.}
\format{A data frame with 19 rows and 2 variables.
\describe{
  \item{height}{variable}
  \item{footlength}{variable}
}}
\source{
\url{http://people.kzoo.edu/enordmoe/math365/data/}
}
\usage{
stmeas
}
\description{
The description of the stmeas data set.
}
\keyword{datasets}
