% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{strikes}
\alias{strikes}
\title{The strikes data set.}
\format{A data frame with 20 rows and 3 variables.
\describe{
  \item{X1}{variable}
  \item{before}{variable}
  \item{after}{variable}
}}
\source{
\url{http://people.kzoo.edu/enordmoe/math365/data/}
}
\usage{
strikes
}
\description{
The description of the strikes data set.
}
\keyword{datasets}
