% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{Titanic}
\alias{Titanic}
\title{The Titanic data set.}
\format{A data frame with 658 rows and 3 variables.
\describe{
  \item{ID}{variable}
  \item{Survived}{variable}
  \item{Age}{variable}
}}
\source{
\url{http://people.kzoo.edu/enordmoe/math365/data/}
}
\usage{
Titanic
}
\description{
The description of the Titanic data set.
}
\keyword{datasets}
