% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{Turbine}
\alias{Turbine}
\title{The Turbine data set.}
\format{A data frame with 168 rows and 4 variables.
\describe{
  \item{Date2010}{variable}
  \item{AveKW}{variable}
  \item{AveSpeed}{variable}
  \item{Production}{variable}
}}
\source{
\url{http://people.kzoo.edu/enordmoe/math365/data/}
}
\usage{
Turbine
}
\description{
The description of the Turbine data set.
}
\keyword{datasets}
