% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{TV}
\alias{TV}
\title{The TV data set.}
\format{A data frame with 20 rows and 3 variables.
\describe{
  \item{ID}{variable}
  \item{Times}{variable}
  \item{Cable}{variable}
}}
\source{
\url{http://people.kzoo.edu/enordmoe/math365/data/}
}
\usage{
TV
}
\description{
The description of the TV data set.
}
\keyword{datasets}
