% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{TXBirths}
\alias{TXBirths}
\title{The TXBirths data set.}
\format{A data frame with 1587 rows and 3 variables.
\describe{
  \item{X1}{variable}
  \item{Smoker}{variable}
  \item{Weight}{variable}
}}
\source{
\url{http://people.kzoo.edu/enordmoe/math365/data/}
}
\usage{
TXBirths
}
\description{
The description of the TXBirths data set.
}
\keyword{datasets}
