% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{TXBirths2004}
\alias{TXBirths2004}
\title{The TXBirths2004 data set.}
\format{A data frame with 1587 rows and 8 variables.
\describe{
  \item{ID}{variable}
  \item{MothersAge}{variable}
  \item{Smoker}{variable}
  \item{Gender}{variable}
  \item{Weight}{variable}
  \item{Gestation}{variable}
  \item{Number}{variable}
  \item{Multiple}{variable}
}}
\source{
\url{http://people.kzoo.edu/enordmoe/math365/data/}
}
\usage{
TXBirths2004
}
\description{
The description of the TXBirths2004 data set.
}
\keyword{datasets}
