% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{Verizon}
\alias{Verizon}
\title{The Verizon data set.}
\format{A data frame with 1687 rows and 2 variables.
\describe{
  \item{Time}{variable}
  \item{Group}{variable}
}}
\source{
\url{http://people.kzoo.edu/enordmoe/math365/data/}
}
\usage{
Verizon
}
\description{
The description of the Verizon data set.
}
\keyword{datasets}
