% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{Volleyball2009}
\alias{Volleyball2009}
\title{The Volleyball2009 data set.}
\format{A data frame with 30 rows and 4 variables.
\describe{
  \item{Team}{variable}
  \item{HitPercent}{variable}
  \item{Assts}{variable}
  \item{Kills}{variable}
}}
\source{
\url{http://people.kzoo.edu/enordmoe/math365/data/}
}
\usage{
Volleyball2009
}
\description{
The description of the Volleyball2009 data set.
}
\keyword{datasets}
