% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{wafers}
\alias{wafers}
\title{The wafers data set.}
\format{A data frame with 25 rows and 2 variables.
\describe{
  \item{Instrument}{variable}
  \item{Resistance}{variable}
}}
\source{
\url{http://people.kzoo.edu/enordmoe/math365/data/}
}
\usage{
wafers
}
\description{
The description of the wafers data set.
}
\keyword{datasets}
