% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{Watertable}
\alias{Watertable}
\title{The Watertable data set.}
\format{A data frame with 360 rows and 2 variables.
\describe{
  \item{Depth}{variable}
  \item{Alive}{variable}
}}
\source{
\url{http://people.kzoo.edu/enordmoe/math365/data/}
}
\usage{
Watertable
}
\description{
The description of the Watertable data set.
}
\keyword{datasets}
